requirements.txt
streamlit
tensorflow
pillow
numpy
pydot
graphviz
